refactor: 인기 종목 상수를 도메인 레이어로 이동
- popular-tickers.ts를 도메인 상수로 이동
- 모든 import 경로 업데이트
- sitemap, not-found, page에서 새 경로로 참조
- 레이어 의존성 규칙 준수
