+++ src/app/not-found.tsx
@@ -4,7 +4,7 @@ import { SITE_NAME } from '@/lib/seo';
 import {
     POPULAR_TICKERS,
     POPULAR_TICKERS_DISPLAY_COUNT,
-} from '@/app/popular-tickers';
+} from '@/domain/constants/popular-tickers';
 import { Footer } from '@/components/layout/Footer';
 import { ContactDialog } from '@/components/layout/ContactDialog';
