+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import { SITE_DESCRIPTION, SITE_NAME, SITE_URL } from '@/lib/seo';
 import {
     POPULAR_TICKERS,
     POPULAR_TICKERS_DISPLAY_COUNT,
-} from '@/app/popular-tickers';
+} from '@/domain/constants/popular-tickers';
 import {
     countSkillFiles,
     FileSkillsLoader,
