+++ src/domain/constants/popular-tickers.ts
